+++ src/main/java/com/magento/idea/magento2plugin/reference/js/JsReferenceContributor.java
@@ -15,6 +15,7 @@ import com.intellij.patterns.PlatformPatterns;
 import com.intellij.psi.PsiReferenceContributor;
 import com.intellij.psi.PsiReferenceRegistrar;
 import com.magento.idea.magento2plugin.reference.provider.FilePathReferenceProvider;
+import com.magento.idea.magento2plugin.reference.provider.KnockoutTemplateReferenceProvider;
 import com.magento.idea.magento2plugin.reference.provider.ModuleNameReferenceProvider;
 import com.magento.idea.magento2plugin.reference.provider.RequireJsPreferenceReferenceProvider;
 import com.magento.idea.magento2plugin.util.RegExUtil;
@@ -49,6 +50,11 @@ public class JsReferenceContributor extends PsiReferenceContributor {
                 JSPatterns.jsLiteralExpression()
                         .withText(string().matches(".*\\W" + RegExUtil.FILE_PATH + ".*")),
                 new RequireJsPreferenceReferenceProvider()
+        );
+
+        registrar.registerReferenceProvider(
+                JSPatterns.jsLiteralExpression(),
+                new KnockoutTemplateReferenceProvider()
         );
     }
 }