+++ src/test/testData/reference/js/KnockoutTemplateReferenceRegistrar/templatePropertyMustHaveReference/component.js
@@ -1,0 +1,9 @@
+define(['uiComponent'], function (Component) {
+    'use strict';
+
+    return Component.extend({
+        defaults: {
+            template: 'Foo_Bar/template/templa<caret>te'
+        }
+    });
+});